Add sesskey protection

+++ index.php
@@ -28,26 +28,37 @@ $PAGE->set_url(new moodle_url('/local/greetings/index.php'));
 $PAGE->set_pagelayout('standard');
 $PAGE->set_title($SITE->fullname);
 $PAGE->set_heading(get_string('pluginname', 'local_greetings'));
+
 require_login();
+
 if (isguestuser()) {
     throw new moodle_exception('noguest');
 }
 
 echo $OUTPUT->header();
 $allowpost = has_capability('local/greetings:postmessages', $context);
+$deletepost = has_capability('local/greetings:deleteownmessage', $context);
 $deleteanypost = has_capability('local/greetings:deleteanymessage', $context);
 
 $action = optional_param('action', '', PARAM_TEXT);
 
 if ($action == 'del') {
+    require_sesskey();
     $id = required_param('id', PARAM_TEXT);
 
     // TODO: Confirm before deleting.
-    $DB->delete_records('local_greetings_messages', array('id' => $id));
+    if ($deleteanypost || $deletepost) {
+        $DB->delete_records('local_greetings_messages', array('id' => $id));
+    }
 }
+
 $messageform = new \local_greetings\form\message_form();
+
 if ($data = $messageform->get_data()) {
     require_capability('local/greetings:postmessages', $context);
+
+    require_sesskey();
+
     $message = required_param('message', PARAM_TEXT);
 
     if (!empty($message)) {
@@ -69,41 +80,41 @@ if ($allowpost) {
     $messageform->display();
 }
 if (has_capability('local/greetings:viewmessages', $context)) {
-$userfields = \core_user\fields::for_name()->with_identity($context);
-$userfieldssql = $userfields->get_sql('u');
+    $userfields = \core_user\fields::for_name()->with_identity($context);
+    $userfieldssql = $userfields->get_sql('u');
 
-$sql = "SELECT m.id, m.message, m.timecreated, m.userid {$userfieldssql->selects}
+    $sql = "SELECT m.id, m.message, m.timecreated, m.userid {$userfieldssql->selects}
         FROM {local_greetings_messages} m
         LEFT JOIN {user} u ON u.id = m.userid
         ORDER BY timecreated DESC";
 
-$messages = $DB->get_records_sql($sql);
-echo $OUTPUT->box_start('card-columns');
+    $messages = $DB->get_records_sql($sql);
+    echo $OUTPUT->box_start('card-columns');
 
-foreach ($messages as $m) {
-    echo html_writer::start_tag('div', array('class' => 'card'));
-    echo html_writer::start_tag('div', array('class' => 'card-body'));
-    echo html_writer::tag('p', format_text($m->message, FORMAT_PLAIN), array('class' => 'card-text'));
-    echo html_writer::tag('p', get_string('postedby', 'local_greetings', $m->firstname), array('class' => 'card-text'));
-    echo html_writer::start_tag('p', array('class' => 'card-text'));
-    echo html_writer::tag('small', userdate($m->timecreated), array('class' => 'text-muted'));
-    echo html_writer::end_tag('p');
+    foreach ($messages as $m) {
+        echo html_writer::start_tag('div', array('class' => 'card'));
+        echo html_writer::start_tag('div', array('class' => 'card-body'));
+        echo html_writer::tag('p', format_text($m->message, FORMAT_PLAIN), array('class' => 'card-text'));
+        echo html_writer::tag('p', get_string('postedby', 'local_greetings', $m->firstname), array('class' => 'card-text'));
+        echo html_writer::start_tag('p', array('class' => 'card-text'));
+        echo html_writer::tag('small', userdate($m->timecreated), array('class' => 'text-muted'));
+        echo html_writer::end_tag('p');
 
-    if ($deleteanypost) {
-        echo html_writer::start_tag('p', array('class' => 'card-footer text-center'));
-        echo html_writer::link(
+        if ($deleteanypost || ($deletepost && $m->userid == $USER->id)) {
+            echo html_writer::start_tag('p', array('class' => 'card-footer text-center'));
+            echo html_writer::link(
             new moodle_url(
                 '/local/greetings/index.php',
-                array('action' => 'del', 'id' => $m->id)
+                array('action' => 'del', 'id' => $m->id, 'sesskey' => sesskey())
             ),
             $OUTPUT->pix_icon('t/delete', '') . get_string('delete')
-        );
-        echo html_writer::end_tag('p');
-    }
+            );
+            echo html_writer::end_tag('p');
+        }
 
-    echo html_writer::end_tag('div');
-    echo html_writer::end_tag('div');
-}
+        echo html_writer::end_tag('div');
+        echo html_writer::end_tag('div');
+    }
 }
 echo $OUTPUT->box_end();
 echo $OUTPUT->footer();
